# docs/swagger.yaml
basePath: /
definitions:
  model.Admin:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      pass:
        type: string
    type: object
  model.Assignment:
    properties:
      exercise:
        $ref: '#/definitions/model.Exercise'
      exerciseId:
        type: integer
      id:
        type: integer
      link:
        type: string
      name:
        type: string
      student:
        $ref: '#/definitions/model.Student'
      studentId:
        type: integer
    type: object
...
  model.Course:
    properties:
      codeCourse:
        type: string
      id:
        type: integer
      lessons:
        type: integer
      name:
        type: string
      subject:
        $ref: '#/definitions/model.Subject'
      subjectId:
        type: integer
      teacher:
        $ref: '#/definitions/model.Teacher'
      teacherId:
        type: integer
    type: object
  model.CourseOfStudent:
    properties:
      course:
        $ref: '#/definitions/model.Course'
      courseId:
        type: integer
      id:
        type: integer
      student:
        $ref: '#/definitions/model.Student'
      studentId:
        type: integer
    type: object
  model.Exercise:
    properties:
      course:
        $ref: '#/definitions/model.Course'
      courseId:
        type: integer
      id:
        type: integer
      link:
        type: string
      name:
        type: string
    type: object
...
  model.Student:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      pass:
        type: string
    type: object
  model.Subject:
    properties:
      codeSubject:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.Teacher:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      pass:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Api canvas project
  title: Canvas Project
  version: "2.0"
paths:
...
  /login/admin:
    post:
      consumes:
      - application/json
      description: Admin login
      parameters:
      - description: Account admin
        format: model.Admin
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Admin'
      produces:
      - application/json
      responses: {}
      tags:
      - Login
  /login/student:
    post:
      consumes:
      - application/json
      description: Login student
      parameters:
      - description: Account student
        format: model.Student
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Student'
      produces:
      - application/json
      responses: {}
      tags:
      - Login
  /login/teacher:
    post:
      consumes:
      - application/json
      parameters:
      - description: Account Teaccher
        format: model.Teacher
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Teacher'
      produces:
      - application/json
      responses: {}
      tags:
      - Login
  /student/joincourse:
    post:
      consumes:
      - application/json
      description: Student join course
      parameters:
      - description: Course of student
        format: model.CourseOfStudent
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.CourseOfStudent'
      produces:
      - application/json
      responses: {}
      tags:
      - Student with course
...
  /student/my_course/{id}:
    get:
      consumes:
      - application/json
      description: Student get course of student by id
      parameters:
      - description: Course of student id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Student with course
  /student/my_course/{id}/exercise:
    get:
      consumes:
      - application/json
      description: Student get exercise of course
      parameters:
      - description: Course Of Student id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Student with exercise
  /student/my_course/{id}/exercise/{exercise_id}:
    get:
      consumes:
      - application/json
      description: Student get exercise of course by id
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      - description: Exercise Id
        in: path
        name: exercise_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Student with exercise
  /student/my_course/{id}/exercise/{exercise_id}/assignment:
    get:
      consumes:
      - application/json
      description: Student get assignment of exercise
      parameters:
      - description: Exercise id
        in: path
        name: exercise_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Student with assignment
  /student/my_course/{id}/exercise/{exercise_id}/assignment_create:
    post:
      consumes:
      - application/json
      description: Student create assignment
      parameters:
      - description: Exercise id
        in: path
        name: exercise_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Student with assignment
  /student/my_course/{id}/exercise/{exercise_id}/assignment_update:
    put:
      consumes:
      - application/json
      description: Student update assignment
      parameters:
      - description: Exercise Id
        in: path
        name: exercise_id
        required: true
        type: string
      - description: Assignment need update
        format: model.Assignment
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Assignment'
      produces:
      - application/json
      responses: {}
      tags:
      - Student with assignment
...
  /teacher/courses/{id}/exercise/{exercise_id}/assignment:
    get:
      consumes:
      - application/json
      description: Teacher get assignment of exercise
      parameters:
      - description: exercise id
        in: path
        name: exercise_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with assignment
  /teacher/courses/{id}/exercise/{exercise_id}/assignment/{assignment_id}:
    get:
      consumes:
      - application/json
      description: Teacher get assignment
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      - description: Exercise id
        in: path
        name: exercise_id
        required: true
        type: string
      - description: Assignment id
        in: path
        name: assignment_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with assignment
  /teacher/courses/{id}/exercise/{exercise_id}/assignment/update:
    put:
      consumes:
      - application/json
      description: Teacher update assignment
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      - description: Exercise id
        in: path
        name: exercise_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with assignment
...
swagger: "2.0"